Fix: js 파일 모듈화 및 eslint, prettier 설정 수정
js 파일들을 잘게 나누어 index.js 한 파일만 script 태그로 읽어와 다른 파일
들은 import 해와서 실행하는 방식으로 구조를 고쳤다.

중간에 파일 경로와 관련된 404 에러가 나서 애를 먹었지만 결국 해결하였다.
import 해올때 확장자명인 '.js'를 적어주지 않아서 일어난 문제였다.

그리고 package.json에 있었던 eslint, prettier 설정들을 따로 파일을 만들어
분리하였다.
