+++ .eslintrc.json
@@ -1,0 +1,23 @@
+{
+  "extends": [
+    "airbnb-base",
+    "plugin:prettier/recommended"
+  ],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "rules": {
+    "object-curly-newline": 0,
+    "prefer-const": 0,
+    "no-param-reassign": 0,
+    "import/prefer-default-export": 0,
+    "import/extensions": 0
+  },
+  "env": {
+    "browser": true,
+    "node": true,
+    "commonjs": true,
+    "es2021": true
+  }
+}
