+++ index.html
@@ -277,7 +277,6 @@
 				</ul>
 			</section>
 		</main>
-		<script src="./src/js/dateInput.js"></script>
-		<script src="./src/js/dropdown.js"></script>
+		<script type="module" src="./src/js/index.js"></script>
 	</body>
 </html>
