+++ src/js/index.js
@@ -1,0 +1,9 @@
+import { initInputDate } from './initInputDate.js';
+import './dropdown.js';
+import './formValidation.js';
+
+const init = () => {
+	initInputDate();
+};
+
+window.onload = init;
