+++ src/js/initInputDate.js
@@ -1,0 +1,8 @@
+import { getToday } from './getToday.js';
+
+const initInputDate = () => {
+	const dateInput = document.querySelector('#inputDate');
+	dateInput.value = getToday();
+};
+
+export { initInputDate };
